third_party/blink/web_tests/fast/xmlhttprequest/xmlhttprequest-responsetype-sync-request-expected.txt

This tests that the XMLHttpRequest responseType attribute is not modifiable for synchronous HTTP(S) requests.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.responseType = 'document'; threw exception InvalidAccessError: Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be changed for synchronous requests made from a document..
PASS xhr.responseType = 'document'; threw exception InvalidAccessError: Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be changed for synchronous requests made from a document..
PASS window.location.protocol is "file:"
PASS xhr.responseType = 'document'; threw exception InvalidAccessError: Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be changed for synchronous requests made from a document..
PASS xhr.responseType = 'document'; threw exception InvalidAccessError: Failed to set the 'responseType' property on 'XMLHttpRequest': The response type cannot be changed for synchronous requests made from a document..
PASS successfullyParsed is true

TEST COMPLETE
